+++ Core/Src/Loadcell.c
@@ -1,0 +1,12 @@
+/*
+ * Loadcell.c
+ *
+ *  Created on: Dec 17, 2020
+ *      Author: EdoOffice
+ */
+
+#include "Loadcell.h"
+
+void loadcell_init() {
+
+}
